requirements.txt
flask
deepface
numpy
opencv-python-headless
azure-storage-blob
pyodbc
